added second NIC to ngnix for Abiquo
