+++ pom.xml
@@ -4,7 +4,7 @@
 
     <groupId>io.cloudsoft.social-apps</groupId>
     <artifactId>brooklyn-social-apps</artifactId>
-    <version>0.1.0-SNAPSHOT</version>
+    <version>0.6.0-SNAPSHOT</version>
     <packaging>jar</packaging>
 
     <name>Brooklyn Social Apps</name>
@@ -28,7 +28,9 @@
             <groupId>io.brooklyn</groupId>
             <artifactId>brooklyn-all</artifactId>
             <version>${brooklyn.version}</version>
+            <classifier>with-dependencies</classifier>
         </dependency>
+        
         <dependency>
             <groupId>org.testng</groupId>
             <artifactId>testng</artifactId>
@@ -41,11 +43,6 @@
             <version>${brooklyn.version}</version>
             <classifier>tests</classifier>
             <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>io.brooklyn</groupId>
-            <artifactId>brooklyn-software-database</artifactId>
-            <version>${brooklyn.version}</version>
         </dependency>
         <dependency>
             <groupId>io.brooklyn</groupId>
